prompt box for deleting an item + UI changes

+++ src/app/posts/post-list/post-list.component.ts
@@ -69,13 +69,17 @@ export class PostListComponent implements OnInit, OnDestroy {
   }
 
   onDelete(postId: string) {
-    this.Loading = true;
-    this.postsService.deletePost(postId).subscribe(() => {
-      this.postsService.getPosts(this.postsPerPage, this.currentPage);
-    }, () => {
-      // fix the loading screen after error dialog
-      this.Loading = false;
-    });
+    if (confirm('Are you sure you want to delete this item' + postId)) {
+      this.Loading = true;
+      this.postsService.deletePost(postId).subscribe(() => {
+        this.postsService.getPosts(this.postsPerPage, this.currentPage);
+      }, () => {
+        // fix the loading screen after error dialog
+        this.Loading = false;
+      });
+    }else {
+      console.log('didnt delete item');
+    }
   }
 
   ngOnDestroy() {
